+++ jbSA-GUI/src/org/ilap/javabacula/model/PoolManagerModel.java
@@ -83,9 +83,9 @@ public class PoolManagerModel  extends BaculaObjectManagerModel {
                         } else if (tokens[i].equals("use_once")) {
                             pool.setUseOnce(tokens[i+1] == "1");
                         } else if (tokens[i].equals("cat_files")) {
-                            pool.setCatFiles(new Integer (tokens[i+1]));
+                            pool.setCatFiles(new Long (tokens[i+1]));
                         } else if (tokens[i].equals("max_vols")) {
-                            pool.setMaxVols(new Integer (tokens[i+1]));
+                            pool.setMaxVols(new Long (tokens[i+1]));
                         } else if (tokens[i].equals("auto_prune")) {
                             pool.setAutoPrune(tokens[i+1] == "1");
                         } else if (tokens[i].equals("VolRetention")) {
@@ -105,15 +105,15 @@ public class PoolManagerModel  extends BaculaObjectManagerModel {
                         } else if (tokens[i].equals("PurgeOldest")) {
                             pool.setPurgeOldest(tokens[i+1] == "1");
                         } else if (tokens[i].equals("MaxVolJobs")) {
-                            pool.setMaxVolJobs(new Integer (tokens[i+1]));
+                            pool.setMaxVolJobs(new Long (tokens[i+1]));
                         } else if (tokens[i].equals("MaxVolFiles")) {
-                            pool.setMaxVolFiles(new Integer (tokens[i+1]));
+                            pool.setMaxVolFiles(new Long (tokens[i+1]));
                         } else if (tokens[i].equals("MigTime")) {
                             pool.setMigTime(tokens[i+1]);
                         } else if (tokens[i].equals("MigHiBytes")) {
-                            pool.setMigHiBytes(new Integer (tokens[i+1]));
+                            pool.setMigHiBytes(new Long (tokens[i+1]));
                         } else if (tokens[i].equals("MigLoBytes")) {
-                            pool.setMigLoBytes(new Integer (tokens[i+1]));
+                            pool.setMigLoBytes(new Long (tokens[i+1]));
                         }
                     } // Even or odd
                 } // for tokens...
