+++ jbSA-GUI/src/org/ilap/javabacula/model/StorageManagerModel.java
@@ -82,7 +82,7 @@ public class StorageManagerModel extends BaculaObjectManagerModel {
                         } else if (tokens[i].equals("SDport")) {
                             std.setPort(new Integer(tokens[i+1]));
                         } else if (tokens[i].equals("MaxJobs")) {
-                            std.setMaxJobs(new Integer(tokens[i+1]));
+                            std.setMaxJobs(new Long(tokens[i+1]));
                         } else if (tokens[i].equals("DeviceName")) {
                             std.setDeviceName(tokens[i+1]);
                         } else if (tokens[i].equals("MediaType")) {
